% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scriptmap.R
\name{condMultiFilter}
\alias{condMultiFilter}
\title{Conditionally filter a data frame by multiple regex's}
\usage{
condMultiFilter(dfr, col, rgx, ...)
}
\arguments{
\item{dfr}{the data frame to be filtered/subsetted}

\item{col}{character: the name of the column of the data frame to
which filtering is to be applied.}

\item{rgx}{a character vector of regular expressions.  Remember
that the character strings are parsed by the R interpreter before
being parsed by the regular expression parser, so backslashes must
be 'doubled up'.}

\item{...}{other arguments to be passed to \code{\link{grepl}}.}
}
\value{
A data frame.
}
\description{
Filter (subset) a data frame by a vector \code{rgx} of regular
expressions applied to a specified column of the data frame.  The
filter is essentially '\code{rgx[1] | rgx[2] | ... rgx[N]}'.
Filtering is conditional, in that if the argument \code{rgx} is
\code{NULL}, the dataframe is returned unchanged.
}
\author{
Alexander Zwart (alec.zwart at csiro.au)
}

